+++ src/marksapp/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
-from .models import Bookmark, Tag
+from .models import Bookmark, Tag, Profile
 
 admin.site.register(Bookmark)
 admin.site.register(Tag)
-# Register your models here.
+admin.site.register(Profile)